Changes
Zellij was the problem when rendering some specific highlights, one of
those was uncercurl, it caused that when using the spelling feature in
Neovim, after the "bad spelled" word the following were the same style,
thin, same colors, etc., so to solve the issue:

1. Disabled the spell feature and made a keymap to toggle it.
2. Created a new file for the "fixes" I didn't know where to save them,
in those fixes are some that were initially in the init.lua.
3. In the fixes included the "pre-stop-using-at-the-moment" Zellij that
were changing the highlights of Spell* to be "underline" instead of
"undercurl".